+++ Cadmus.Export.Test/Filters/Iso639FilterTest.cs
@@ -1,4 +1,5 @@
 ﻿using Cadmus.Export.Filters;
+using System.Threading.Tasks;
 using Xunit;
 
 namespace Cadmus.Export.Test.Filters;
@@ -6,40 +7,40 @@ namespace Cadmus.Export.Test.Filters;
 public sealed class Iso639FilterTest
 {
     [Fact]
-    public void Apply_NoMatch_Unchanged()
+    public async Task Apply_NoMatch_Unchanged()
     {
         Iso639TextFilter filter = new();
         const string text = "Hello, world!";
 
-        string? filtered = filter.Apply(text)?.ToString();
+        string? filtered = (await filter.ApplyAsync(text))?.ToString();
 
         Assert.Equal(text, filtered);
     }
 
     [Fact]
-    public void Apply_MatchInvalidCode_Code()
+    public async Task Apply_MatchInvalidCode_Code()
     {
         Iso639TextFilter filter = new();
         const string text = "Hello, ^^xyz world!";
 
-        string? filtered = filter.Apply(text)?.ToString();
+        string? filtered = (await filter.ApplyAsync(text))?.ToString();
 
         Assert.Equal("Hello, xyz world!", filtered);
     }
 
     [Fact]
-    public void Apply_Match_Changed()
+    public async Task Apply_Match_Changed()
     {
         Iso639TextFilter filter = new();
         const string text = "Hello, ^^eng and ^^ita world!";
 
-        string? filtered = filter.Apply(text)?.ToString();
+        string? filtered = (await filter.ApplyAsync(text))?.ToString();
 
         Assert.Equal("Hello, English and Italian world!", filtered);
     }
 
     [Fact]
-    public void Apply_Match2Letters_Changed()
+    public async Task Apply_Match2Letters_Changed()
     {
         Iso639TextFilter filter = new();
         filter.Configure(new Iso639FilterOptions
@@ -49,7 +50,7 @@ public sealed class Iso639FilterTest
         });
         const string text = "Hello, ^^en and ^^it world!";
 
-        string? filtered = filter.Apply(text)?.ToString();
+        string? filtered = (await filter.ApplyAsync(text))?.ToString();
 
         Assert.Equal("Hello, English and Italian world!", filtered);
     }