+++ Cadmus.Export.Test/Filters/MarkdownRendererFilterTest.cs
@@ -1,4 +1,5 @@
 ﻿using Cadmus.Export.Filters;
+using System.Threading.Tasks;
 using Xunit;
 
 namespace Cadmus.Export.Test.Filters;
@@ -18,27 +19,28 @@ public sealed class MarkdownRendererFilterTest
     }
 
     [Fact]
-    public void Apply_NoRegion_Unchanged()
+    public async Task Apply_NoRegion_Unchanged()
     {
         MarkdownTextFilter filter = GetFilter();
 
-        string? result = filter.Apply("No markdown here")?.ToString();
+        string? result = (await filter.ApplyAsync("No markdown here"))?.ToString();
 
         Assert.Equal("No markdown here", result);
     }
 
     [Fact]
-    public void Apply_Regions_Ok()
+    public async Task Apply_Regions_Ok()
     {
         MarkdownTextFilter filter = GetFilter();
 
-        string? result = filter.Apply("Hello. <_md>This *is* MD.</_md> End.")?.ToString();
+        string? result = (await filter.ApplyAsync(
+            "Hello. <_md>This *is* MD.</_md> End."))?.ToString();
 
         Assert.Equal("Hello. <p>This <em>is</em> MD.</p>\n End.", result);
     }
 
     [Fact]
-    public void Apply_WholeText_Ok()
+    public async Task Apply_WholeText_Ok()
     {
         MarkdownTextFilter filter = new();
         filter.Configure(new MarkdownRendererFilterOptions
@@ -46,7 +48,7 @@ public sealed class MarkdownRendererFilterTest
             Format = "html"
         });
 
-        string? result = filter.Apply("This *is* MD.")?.ToString();
+        string? result = (await filter.ApplyAsync("This *is* MD."))?.ToString();
 
         Assert.Equal("<p>This <em>is</em> MD.</p>\n", result);
     }