+++ Cadmus.Export.Test/Filters/SentenceSplitRendererFilterTest.cs
@@ -1,4 +1,5 @@
 ﻿using Cadmus.Export.Filters;
+using System.Threading.Tasks;
 using Xunit;
 
 namespace Cadmus.Export.Test.Filters;
@@ -21,82 +22,85 @@ public sealed class SentenceSplitRendererFilterTest
     }
 
     [Fact]
-    public void Apply_NoMarker_AppendedNL()
+    public async Task Apply_NoMarker_AppendedNL()
     {
         SentenceSplitTextFilter filter = GetFilter();
 
-        string? result = filter.Apply("Hello, world")?.ToString();
+        string? result = (await filter.ApplyAsync("Hello, world"))?.ToString();
 
         Assert.Equal("Hello, world\n", result);
     }
 
     [Fact]
-    public void Apply_InitialMarker_AppendedNL()
+    public async Task Apply_InitialMarker_AppendedNL()
     {
         SentenceSplitTextFilter filter = GetFilter();
 
-        string? result = filter.Apply("!Hello, world")?.ToString();
+        string? result = (await filter.ApplyAsync("!Hello, world"))?.ToString();
 
         Assert.Equal("!Hello, world\n", result);
     }
 
     [Fact]
-    public void Apply_MarkerSequence_AppendedNL()
+    public async Task Apply_MarkerSequence_AppendedNL()
     {
         SentenceSplitTextFilter filter = GetFilter(true);
 
-        string? result = filter.Apply("Hello... world?!")?.ToString();
+        string? result = (await filter.ApplyAsync("Hello... world?!"))?.ToString();
 
         Assert.Equal("Hello...\nworld?!\n", result);
     }
 
     [Fact]
-    public void Apply_Markers_Split()
+    public async Task Apply_Markers_Split()
     {
         SentenceSplitTextFilter filter = GetFilter();
 
-        string? result = filter.Apply("Hello! I am world.")?.ToString();
+        string? result = (await filter.ApplyAsync("Hello! I am world."))?.ToString();
 
         Assert.Equal("Hello! \nI am world.\n", result);
     }
 
     [Fact]
-    public void Apply_MarkersWithBlack_Split()
+    public async Task Apply_MarkersWithBlack_Split()
     {
         SentenceSplitTextFilter filter = GetFilter(true);
 
-        string? result = filter.Apply("Hello (can you believe?) world! End.")
-            ?.ToString();
+        string? result = (await filter.ApplyAsync(
+            "Hello (can you believe?) world! End."))?.ToString();
 
         Assert.Equal("Hello (can you believe?) world!\nEnd.\n", result);
     }
 
     [Fact]
-    public void Apply_MarkersWithTrim_Split()
+    public async Task Apply_MarkersWithTrim_Split()
     {
         SentenceSplitTextFilter filter = GetFilter(true);
 
-        string? result = filter.Apply("Hello! I am world.")?.ToString();
+        string? result = (await filter.ApplyAsync(
+            "Hello! I am world."))?.ToString();
 
         Assert.Equal("Hello!\nI am world.\n", result);
     }
 
     [Fact]
-    public void Apply_MarkersWithCrLf_Split()
+    public async Task Apply_MarkersWithCrLf_Split()
     {
         SentenceSplitTextFilter filter = GetFilter();
 
-        string? result = filter.Apply("Hello! I\r\nam world.")?.ToString();
+        string? result = (await filter.ApplyAsync(
+            "Hello! I\r\nam world."))?.ToString();
 
         Assert.Equal("Hello! \nI am world.\n", result);
     }
 
     [Fact]
-    public void Apply_MarkersWithCrLfTrim_Split()
+    public async Task Apply_MarkersWithCrLfTrim_Split()
     {
         SentenceSplitTextFilter filter = GetFilter(true);
 
-        string? result = filter.Apply("Hello! I\r\nam world.")?.ToString();
+        string? result = (await filter.ApplyAsync(
+            "Hello! I\r\nam world."))?.ToString();
 
         Assert.Equal("Hello!\nI am world.\n", result);
     }