+++ Cadmus.Export/Filters/Iso639TextFilter.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Reflection;
 using System.Text;
 using System.Text.RegularExpressions;
+using System.Threading.Tasks;
 
 namespace Cadmus.Export.Filters;
 
@@ -82,7 +83,7 @@ public sealed class Iso639TextFilter : TextFilter<string>,
     /// <param name="text">The text.</param>
     /// <param name="context">The optional context.</param>
     /// <returns>Filtered text or null.</returns>
-    protected override object? DoApply(string? text,
+    protected override async Task<object?> DoApplyAsync(string? text,
         IHasDataDictionary? context = null)
     {
         if (string.IsNullOrEmpty(text) || _isoRegex == null) return text;