+++ Cadmus.Export/Filters/MarkdownTextFilter.cs
@@ -3,6 +3,7 @@ using Fusi.Tools.Configuration;
 using Markdig;
 using Proteus.Core.Text;
 using System;
+using System.Threading.Tasks;
 
 namespace Cadmus.Export.Filters;
 
@@ -34,7 +35,7 @@ public sealed class MarkdownTextFilter : TextFilter<string>,
     /// <param name="text">The text.</param>
     /// <param name="context">The optional context.</param>
     /// <returns>Filtered text or null.</returns>
-    protected override object? DoApply(string? text,
+    protected override async Task<object?> DoApplyAsync(string? text,
         IHasDataDictionary? context = null)
     {
         if (string.IsNullOrEmpty(text) || _options == null) return text;