+++ Cadmus.Export/Filters/MongoTokenExtractorTextFilter.cs
@@ -10,6 +10,7 @@ using Cadmus.Core.Layers;
 using Fusi.Tools.Text;
 using Proteus.Core.Text;
 using Fusi.Tools;
+using System.Threading.Tasks;
 
 namespace Cadmus.Export.Filters;
 
@@ -60,7 +61,7 @@ public sealed class MongoTokenExtractorTextFilter : TextFilter<string>,
     /// <param name="text">The text.</param>
     /// <param name="context">The optional context.</param>
     /// <returns>Filtered text or null.</returns>
-    protected override object? DoApply(string? text,
+    protected override async Task<object?> DoApplyAsync(string? text,
         IHasDataDictionary? context = null)
     {
         if (_locRegex == null || string.IsNullOrEmpty(text)) return text;