+++ Cadmus.Export/Filters/SourceIdTextFilter.cs
@@ -3,6 +3,7 @@ using Fusi.Tools.Configuration;
 using Proteus.Core.Text;
 using System;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace Cadmus.Export.Filters;
 
@@ -58,7 +59,7 @@ public sealed class SourceIdTextFilter : TextFilter<string>,
     /// <param name="text">The text.</param>
     /// <param name="context">The optional context.</param>
     /// <returns>Filtered text or null.</returns>
-    protected override object? DoApply(string? text,
+    protected override async Task<object?> DoApplyAsync(string? text,
         IHasDataDictionary? context = null)
     {
         if (string.IsNullOrEmpty(text) || context == null) return text;