+++ Proteus.Rendering/ITextTreeFilter.cs
@@ -1,6 +1,5 @@
 ﻿using Fusi.Tools;
 using Fusi.Tools.Data;
-using Proteus.Core;
 
 namespace Proteus.Rendering;